move mock relayer to spec.ts
