+++ libs/ethers/src/lib/relayer/relayer.ts
@@ -16,36 +16,4 @@ export interface IRelayer {
 }
 
 @Injectable({ providedIn: 'root' })
-export class MockRelayer implements IRelayer {
-  async create(name: string, key: string) {
-    return { message: 'mock erc1077 created' };
-  }
-  async prepare(name: string) {
-    return { message: 'mock erc1077 prepare' };
-  }
-  async send(
-    name: string,
-    tx: providers.TransactionRequest
-  ): Promise<providers.TransactionResponse> {
-    const receipt: providers.TransactionReceipt = {
-      byzantium: false
-    };
-    return {
-      confirmations: 0,
-      from: '0x0',
-      wait: async () => receipt,
-      nonce: 0,
-      gasLimit: new BigNumber(0),
-      gasPrice: new BigNumber(0),
-      data: '0x0',
-      value: new BigNumber(0),
-      chainId: 0
-    };
-  }
-  async addKey(name: string, key: string) {
-    return { message: 'mock erc1077 addKey' };
-  }
-  async removeKey(name: string, key: string) {
-    return { message: 'mock erc1077 removeKey' };
-  }
-}
+
